Part 4
We already have a shared GameObject class. It's time to create classes for specific types of game objects.
4.1. Some game objects can move (player and boxes) and some can't (walls and storage locations).
4.1.1. Add the Movable interface to the model package.
4.1.2. The Movable interface must have a void move(int x, int y) method.
4.2. Add a Direction enum to the model package. It must have the following values:
LEFT, RIGHT, UP, and DOWN. This type will be used to describe the direction of an object's movement.
4.3. The "storage location" objects don't support collision detection (the player and boxes can freely move over them). As for other objects, they must not pass through one another, i.e. they should collide. For example, a box can't be pushed through a wall.
4.3.1. Add an abstract CollisionObject class to the model package.
3.4.2. The CollisionObject class must inherit GameObject.
4.3.3. Add the following to the CollisionObject class:
4.3.3.1. A constructor that accepts int x and int y arguments.
4.3.3.2. A boolean isCollision(GameObject gameObject, Direction direction) method.
This method should return true if the current object will collide with the gameObject parameter after moving a distance of BOARD_CELL_SIZE in the direction specified by the direction parameter. Otherwise, it should return false. If (x, y) coordinates match, you have a collision.