Part 5
Box and Player game objects can not only collide with other objects, but also move around the board.
5.1. Create a Player class and a Box class in the model package. Make each of them inherit the most appropriate class.
5.2. The created classes must support the interface responsible for object movement.
5.3. Add constructors that accept int x and int y arguments.
5.4. In each class, implement the method responsible for movement. It must shift the object's (x, y) coordinates by the passed values.
5.5. In each class, implement the method responsible for drawing. This method should: set a color and draw some basic shapes. Be sure that the coordinates of the shape's center are (x, y), and that its width and height match the width and height fields.

Hint: you can draw the player as a solid yellow circle, and boxes — as orange squares with diagonals. But these are just examples. You can choose the colors yourself. The appearance of your objects is limited only by your imagination and the methods available in the Graphics class.