Part 6
It's time to create the StorageLocation class. This class will be responsible for the game board cells that all the boxes must be moved to. Objects of this type must not move across the board or collide with other game objects.
6.1. Add a StorageLocation class to the model package.
6.2. The class must inherit the most appropriate base class.
6.3. Add a constructor that accepts int x and int y arguments. The width and height of storage locations must be equal to 2.
6.4. Implement the method for drawing a storage location.

Hint: you can use red and an unfilled circle.